changing from vue-analytics to vue-gtag and making other necessary changes
