+++ package.json
@@ -20,6 +20,7 @@
     "tailwindcss": "^3.4.10",
     "vue": "latest",
     "vue-cookies": "^1.8.4",
+    "vue-gtag": "^2.0.1",
     "vue-gtag-next": "^1.14.0"
   },
   "devDependencies": {
